removed stats about page
