+++ package.json
@@ -57,6 +57,7 @@
     "plex-api": "^5.3.2",
     "react": "18.2.0",
     "react-dom": "18.3.1",
+    "react-markdown": "^9.0.1",
     "react-select": "^5.8.0",
     "react-toast-notifications": "^2.5.1",
     "react-transition-group": "^4.4.5",
