+++ ui/src/components/Settings/index.tsx
@@ -51,16 +51,6 @@ const SettingsWrapper: React.FC<{ children?: ReactNode }> = (props: {
       route: '/settings/about',
       regex: /^\/settings(\/about)?$/,
     },
-    {
-      text: 'About-ALT',
-      route: '/settings/about-ALT',
-      regex: /^\/settings(\/about-ALT)?$/,
-    },
-    {
-      text: 'Testing',
-      route: '/settings/Testing',
-      regex: /^\/settings(\/Testing)?$/,
-    },
   ]
 
   useEffect(() => {
