--- ui/src/pages/settings/about-ALT/index.tsx
@@ -1,11 +1,0 @@
-import { NextPage } from 'next'
-import SettingsWrapper from '../../../components/Settings'
-import AboutSettings from '../../../components/Settings/About-ALT'
-const SettingsAbout: NextPage = () => {
-  return (
-    <SettingsWrapper>
-      <AboutSettings />
-    </SettingsWrapper>
-  )
-}
-export default SettingsAbout
